refactor(fsm): extract pre-pr inspection runner
